% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netEMD_org.R
\name{net_EMD.IS.org}
\alias{net_EMD.IS.org}
\title{This function calculates the Edgewise Multiplex Depth for network data with importance score matrix.}
\usage{
net_EMD.IS.org(data, w)
}
\arguments{
\item{data}{an n \* N \* N array, the ensember of network, where n is the number of networks, and N is the number of vertex of each network;}

\item{w}{an N\*N importance score matrix, corresponding to the weight of each edge in the network samples.}
}
\value{
A vector of the estimated Edgewise Multiplex Depth for the n network samples
}
\description{
This function calculates the Edgewise Multiplex Depth for network data with importance score matrix.
}
