% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netEMD_org.R
\name{net_EMD.org}
\alias{net_EMD.org}
\title{This function calculates the Edgewise Multiplex Depth for network data.}
\usage{
net_EMD.org(data)
}
\arguments{
\item{data}{an n \* N \* N array, the ensember of network, where n is the number of networks, and N is the number of vertex of each network;}
}
\value{
A vector of the estimated Edgewise Multiplex Depth for the n network samples
}
\description{
This function calculates the Edgewise Multiplex Depth for network data.
}
