% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netEMD.R
\name{net_EMD}
\alias{net_EMD}
\title{This function calculates the Edgewise Multiplex Depth for network data.}
\usage{
net_EMD(data, K = 2, unweighted = F)
}
\arguments{
\item{data}{an n \* N \* N array, the ensember of network, where n is the number of networks, and N is the number of vertex of each network;}

\item{K}{an integer, the parameter for depth.}

\item{unweighted}{a bool, whether the networks are unweighted or weighted.}
}
\value{
A vector of the estimated Edgewise Multiplex Depth for the n network samples.
}
\description{
This function calculates the Edgewise Multiplex Depth for network data.
}
